Moved karpenter components into their own module. Updated versions to, maybe, be managed by renovate.
