+++ src/bridge/lib/versions.py
@@ -29,6 +29,8 @@ EFS_CSI_DRIVER_VERSION = "v2.1.6-eksbuild.1"
 EXTERNAL_DNS_CHART_VERSION = "1.15.0"
 # renovate: datasource=github-releases depName=gateway-api packageName=kubernetes-sigs/gateway-api
 GATEWAY_API_VERSION = "v1.2.1"
+# renovate: datasource=helm depName=karpenter packageName=karpenter registryUrl=oci://public.ecr.aws/karpenter
+KARPENTER_CHART_VERSION = "1.3.2"
 # renovate: datasource=helm depName=traefik packageName=traefik registryUrl=https://traefik.github.io/charts
 TRAEFIK_CHART_VERSION = "34.5.0"
 # renovate: datasource=helm depName=vantage-kubernetes-agent packageName=vantage-kubernetes-agent registryUrl=https://vantage-sh.github.io/helm-charts
